// src/api/client.ts
export class StageData implements IStageData {

    constructor(data?: IStageData) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
    }

    static fromJS(data: any): StageData {
        data = typeof data === 'object' ? data : {};
        let result = new StageData();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        return data;
    }
}

export interface IStageData {
}
// ...
export type BusinessLine = "Architecture" | "Land" | "Transportation";
// ...
export type ChargeType = "Regular" | "Promotional";
// ...
export class NewAdditionalTask implements INewAdditionalTask {
    businessLine!: BusinessLine;
    organization!: string;
    taskDescription!: string;

    constructor(data?: INewAdditionalTask) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.businessLine = _data["businessLine"];
            this.organization = _data["organization"];
            this.taskDescription = _data["taskDescription"];
        }
    }

    static fromJS(data: any): NewAdditionalTask {
        data = typeof data === 'object' ? data : {};
        let result = new NewAdditionalTask();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["businessLine"] = this.businessLine;
        data["organization"] = this.organization;
        data["taskDescription"] = this.taskDescription;
        return data;
    }
}

export interface INewAdditionalTask {
    businessLine: BusinessLine;
    organization: string;
    taskDescription: string;
}

export class NewProjectRequest implements INewProjectRequest {
    projectName!: string;
    organization!: string;
    businessLine!: BusinessLine;
    requesterEmail!: string;
    optionalManualProjectNumber?: string | undefined;
    chargeType!: ChargeType;
    primaryClient!: string;
    additionalTasks!: NewAdditionalTask[];
    stage!: StageKind;
    stageData!: StageData;

    constructor(data?: INewProjectRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
        if (!data) {
            this.additionalTasks = [];
            this.stageData = new StageData();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.projectName = _data["projectName"];
            this.organization = _data["organization"];
            this.businessLine = _data["businessLine"];
            this.requesterEmail = _data["requesterEmail"];
            this.optionalManualProjectNumber = _data["optionalManualProjectNumber"];
            this.chargeType = _data["chargeType"];
            this.primaryClient = _data["primaryClient"];
            if (Array.isArray(_data["additionalTasks"])) {
                this.additionalTasks = [] as any;
                for (let item of _data["additionalTasks"])
                    this.additionalTasks!.push(NewAdditionalTask.fromJS(item));
            }
            this.stage = _data["stage"];
            this.stageData = _data["stageData"] ? StageData.fromJS(_data["stageData"]) : new StageData();
        }
    }

    static fromJS(data: any): NewProjectRequest {
        data = typeof data === 'object' ? data : {};
        let result = new NewProjectRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["projectName"] = this.projectName;
        data["organization"] = this.organization;
        data["businessLine"] = this.businessLine;
        data["requesterEmail"] = this.requesterEmail;
        data["optionalManualProjectNumber"] = this.optionalManualProjectNumber;
        data["chargeType"] = this.chargeType;
        data["primaryClient"] = this.primaryClient;
        if (Array.isArray(this.additionalTasks)) {
            data["additionalTasks"] = [];
            for (let item of this.additionalTasks)
                data["additionalTasks"].push(item ? item.toJSON() : undefined as any);
        }
        data["stage"] = this.stage;
        data["stageData"] = this.stageData ? this.stageData.toJSON() : undefined as any;
        return data;
    }
}

export interface INewProjectRequest {
    projectName: string;
    organization: string;
    businessLine: BusinessLine;
    requesterEmail: string;
    optionalManualProjectNumber?: string | undefined;
    chargeType: ChargeType;
    primaryClient: string;
    additionalTasks: NewAdditionalTask[];
    stage: StageKind;
    stageData: StageData;
}
// ...
export type StageKind = "LONG_LEAD" | "SOURCES_SOUGHT" | "SOLICITATION_RFP_RFQ" | "SUBMITTED" | "SHORTLISTED" | "INTERVIEWED" | "WON_AWAITING_CONTRACT" | "WON_CONTRACT_RECIEVED" | "LOST" | "NO_GO" | "CANCELLED";